+++ examples/js/selection/demo.js
@@ -6,6 +6,7 @@ import UnSelectableTable from './unselectable-table';
 import ClickToSelectTable from './click-to-select-table';
 import DefaultSelectTable from './default-select-table';
 import SelectBgColorTable from './select-bgcolor-table';
+import SelectBgColorDynamicTable from './select-bgcolor-dynamic-table';
 import SelectHookTable from './select-hook-table';
 import HideSelectionColumnTable from './hide-selection-col-table';
 import SelectValidationTable from './select-validation-table';
@@ -80,6 +81,15 @@ class Demo extends React.Component {
             <div className='panel-body'>
               <h5>Source in /examples/js/selection/select-bgcolor-table.js</h5>
               <SelectBgColorTable />
+            </div>
+          </div>
+        </div>
+        <div className='col-md-offset-1 col-md-8'>
+          <div className='panel panel-default'>
+            <div className='panel-heading'>Selected Row Bgcolor Depends on Data Example</div>
+            <div className='panel-body'>
+              <h5>Source in /examples/js/selection/select-bgcolor-dynamic-table.js</h5>
+              <SelectBgColorDynamicTable />
             </div>
           </div>
         </div>